+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Unreleased
+
+## Enhancements
+
+* Adds Beta parameter `Overridable` for OPA `policy set` update API (`PolicySetUpdateOptions`) @mrinalirao [#594](https://github.com/hashicorp/go-tfe/pull/594)
+* Adds new task stage status values representing `canceled`, `errored`, `unreachable` @mrinalirao [#594](https://github.com/hashicorp/go-tfe/pull/594)
+
 # v1.13.0
 
 ## Bug Fixes